Convert source files to UTF-8
Qt 5 assumes that source files are encoded in UTF-8. In particular, string
literals are now decoded as UTF-8 rather than the system's default locale.

Alarmz.rc has not been converted, because the resource compiler, by default,
uses the system's default locale, whether the project is compiled with MSVC or
MinGW.