Stop using QFSFileEngine
The QFSFileEngine class has been made private in Qt 5. It was used to determine
whether the file system was case sensitive or not. This affected how the Options
dialog found the available .qm files in order to populate the list of available
languages for the application. Now, .qm files will always be searched using a
case insensitive comparison.